+++ app-admin/src/app/features/templates/routes.ts
@@ -1,0 +1,16 @@
+import { Routes } from "@angular/router";
+import { getPathComponent } from "@app/shared/functions";
+import { TemplateHomeComponent } from "./template-home/template-home.component";
+import { PromiseHomeComponent } from "./promise-home/promise-home.component";
+import { TemplateGuideComponent } from "./template-guide/template-guide.component";
+
+export const routes: Routes = [
+	{
+		path: '',
+		pathMatch: 'full',
+		redirectTo: '',
+	},
+	...getPathComponent(['', 'home'], TemplateHomeComponent),
+	...getPathComponent(['promises'], PromiseHomeComponent),
+	...getPathComponent(['guide'], TemplateGuideComponent),
+];